Create Day 12: Inheritance.cs
This is my original code...it passed 7/8 test cases in Hankerrank, but inside Visual Studio it came out with the correct results with the same input as the last test case. So, I'm not sure what to make of that. I had to post a different answer to pass hackerrank.